must floor to int

+++ utils/data_loader.py
@@ -62,7 +62,7 @@ class DataLoader:
                     c2.append((ret.x[i], ret.y[i]))
 
             # maj_len = int(max(len(c0), len(c1), len(c2)) / 3)
-            maj_len = len(c2) / 3 * 2
+            maj_len = int(len(c2) / 3 * 2)
             # c0 = resample(c0, n_samples=maj_len, random_state=FLAGS.random_state)
             c1 = resample(c1, n_samples=maj_len, random_state=FLAGS.random_state)
             # c2 = resample(c2, n_samples=maj_len, random_state=FLAGS.random_state)
